final refactor, done.
