+++ plugins/talk-plugin-auth/client/profile-settings/components/ChangePassword.css
@@ -13,22 +13,6 @@
   }
 }
 
-.detailItemContent {
-  min-width: 280px;
-}
-
-.detailItemMessage {
-  flex-grow: 1;
-  display: flex;
-  align-items: center;
-  padding-left: 2px;
-  padding-top: 16px;
-
-  .warningIcon, .checkIcon {
-    font-size: 17px;
-  }
-}
-
 .actions {
   position: absolute;
   top: 10px;
@@ -41,39 +25,6 @@
 .title {
   color: #202020;
   margin: 0 0 20px;
-}
-  
-.detailList {
-  padding: 0;
-  margin: 0;
-  list-style: none;
-}
-
-.detailLabel {
-  color: #4C4C4D;
-  font-size: 1em;
-  display: block;
-  margin-bottom: 4px;
-}
-
-.detailValue {
-  padding: 6px 2px;
-  border: solid 1px #979797;
-  display: block;
-  font-size: 1.1em;
-  border-radius: 2px;
-  background-color: #ffffff;
-  color: #979797;
-  box-sizing: border-box;
-  width: 100%;
-}
-  
-.detailItem {
-  margin-bottom: 12px;
-}
-
-.detailItemContainer {
-  display: flex;
 }
 
 .detailBottomBox {
@@ -90,20 +41,6 @@
   &:hover {
     cursor: pointer;
   }
-}
-
-.checkIcon {
-  color: #00CD73;
-}
-
-.warningIcon {
-  color: #FA4643;
-}
-
-.errorMsg {
-  color: #FA4643;
-  padding-left: 4px;
-  font-size: 0.9em;
 }
 
 .button {
@@ -126,6 +63,16 @@
   &:hover {
     background-color: #399ee2;
     color: white;
+  }
+
+  &:disabled {
+    border-color: #e0e0e0;
+
+    &:hover {
+      background-color: #e0e0e0;
+      color: #4f5c67;
+      cursor: default;
+    }
   }
 }
