+++ plugins/talk-plugin-auth/client/profile-settings/components/ErrorMessage.css
@@ -1,0 +1,9 @@
+.errorMsg {
+  color: #FA4643;
+  padding-left: 4px;
+  font-size: 0.9em;
+}
+
+.warningIcon {
+  color: #FA4643;
+}
