+++ plugins/talk-plugin-auth/client/profile-settings/components/Form.css
@@ -1,0 +1,5 @@
+.detailList {
+  padding: 0;
+  margin: 0;
+  list-style: none;
+}
