+++ plugins/talk-plugin-auth/client/profile-settings/components/InputField.css
@@ -1,0 +1,51 @@
+  
+.detailItem {
+  margin-bottom: 12px;
+}
+
+.detailItemContainer {
+  display: flex;
+}
+
+.detailItemContent {
+  min-width: 280px;
+}
+
+.detailLabel {
+  color: #4C4C4D;
+  font-size: 1em;
+  display: block;
+  margin-bottom: 4px;
+}
+
+.detailValue {
+  padding: 6px 2px;
+  border: solid 1px #979797;
+  display: block;
+  font-size: 1.1em;
+  border-radius: 2px;
+  background-color: #ffffff;
+  color: #979797;
+  box-sizing: border-box;
+  width: 100%;
+}
+
+.detailItemMessage {
+  flex-grow: 1;
+  display: flex;
+  align-items: center;
+  padding-left: 2px;
+  padding-top: 16px;
+
+  .warningIcon, .checkIcon {
+    font-size: 17px;
+  }
+}
+
+.checkIcon {
+  color: #00CD73;
+}
+
+.warningIcon {
+  color: #FA4643;
+}
